+++ packages/core/dlight/src/types/model.d.ts
@@ -1,26 +1,32 @@
 import { ContentKeyName, ContentProp } from "./compTag"
 
-type RemoveDLightLifecycle<T> = Omit<
+type RemoveDLightInternal<T, Props> = Omit<
   T,
-  "willMount" | "didMount" | "didUpdate" | "willUnmount"
+  "willMount" | "didMount" | "didUpdate" | "willUnmount" | keyof Props
 >
 
-export type Modeling<Model, Props = object> = {
-  model: RemoveDLightLifecycle<Model>
-  props: keyof Props extends never
-    ? never
-    : ContentKeyName<Props> extends undefined
-      ? Props
-      : Omit<Props, ContentKeyName<Props>>
-  content: ContentKeyName<Props> extends undefined
-    ? never
-    : Props[ContentKeyName<Props>] extends ContentProp<infer U>
-      ? U
-      : never
-}
+export type Modeling<Model, Props = object> = (props: Props) => Model
 
-export const use: <M extends Modeling<any, any>>(
+type GetProps<T> = keyof T extends never
+  ? never
+  : ContentKeyName<T> extends undefined
+    ? T
+    : Omit<T, ContentKeyName<T>>
+
+type GetContent<T> = ContentKeyName<T> extends undefined
+  ? never
+  : T[ContentKeyName<T>] extends ContentProp<infer U>
+    ? U
+    : never
+
+export const use: <M>(
   model: M,
-  props?: M["props"],
-  content?: M["content"]
-) => M["model"]
+  // @ts-expect-error Model should be a function
+  props?: GetProps<Parameters<M>[0]>,
+  // @ts-expect-error Model should be a function
+  content?: GetContent<Parameters<M>[0]>
+  // @ts-expect-error Model should be a function
+) => RemoveDLightInternal<ReturnType<M>, Parameters<M>[0]>
+
+// @ts-expect-error Model should be a function
+export type ModelType<T> = RemoveDLightInternal<ReturnType<T>, Parameters<T>[0]>