+++ packages/core/transpiler/view-generator/src/NodeGenerators/IfGenerator.ts
@@ -59,7 +59,12 @@ export default class IfGenerator extends CondGenerator {
       })
     ) {
       branches.push({
-        condition: { value: this.t.booleanLiteral(true) },
+        condition: {
+          value: this.t.booleanLiteral(true),
+          dependencyIndexArr: [],
+          dependenciesNode: this.t.arrayExpression([]),
+          dynamic: false,
+        },
         children: [],
       })
     }