+++ packages/core/transpiler/view-generator/src/NodeGenerators/TextGenerator.ts
@@ -8,12 +8,14 @@ export default class TextGenerator extends BaseGenerator {
 
     const dlNodeName = this.generateNodeName()
 
-    this.addInitStatement(this.declareTextNode(dlNodeName, content.value))
+    this.addInitStatement(
+      this.declareTextNode(dlNodeName, content.value, content.dependenciesNode)
+    )
 
-    if (content.dependencyIndexArr && content.dependencyIndexArr.length > 0) {
+    if (content.dynamic) {
       this.addUpdateStatements(
         content.dependencyIndexArr,
-        this.updateTextNode(dlNodeName, content.value)
+        this.updateTextNode(dlNodeName, content.value, content.dependenciesNode)
       )
     }
 
@@ -22,11 +24,12 @@ export default class TextGenerator extends BaseGenerator {
 
   /**
    * @View
-   * ${dlNodeName} = createTextNode(${value})
+   * ${dlNodeName} = createTextNode(${value}, ${deps})
    */
   private declareTextNode(
     dlNodeName: string,
-    value: t.Expression
+    value: t.Expression,
+    dependenciesNode: t.Expression
   ): t.Statement {
     return this.t.expressionStatement(
       this.t.assignmentExpression(
@@ -34,7 +37,7 @@ export default class TextGenerator extends BaseGenerator {
         this.t.identifier(dlNodeName),
         this.t.callExpression(
           this.t.identifier(this.importMap.createTextNode),
-          [value]
+          [value, dependenciesNode]
         )
       )
     )
@@ -42,16 +45,21 @@ export default class TextGenerator extends BaseGenerator {
 
   /**
    * @View
-   * ${dlNodeName} && updateText(${dlNodeName}, ${value})
+   * ${dlNodeName} && updateText(${dlNodeName}, () => ${value}, ${deps})
    */
-  private updateTextNode(dlNodeName: string, value: t.Expression): t.Statement {
+  private updateTextNode(
+    dlNodeName: string,
+    value: t.Expression,
+    dependenciesNode: t.Expression
+  ): t.Statement {
     return this.t.expressionStatement(
       this.t.logicalExpression(
         "&&",
         this.t.identifier(dlNodeName),
         this.t.callExpression(this.t.identifier(this.importMap.updateText), [
           this.t.identifier(dlNodeName),
-          value,
+          this.t.arrowFunctionExpression([], value),
+          dependenciesNode,
         ])
       )
     )