+++ packages/tools/vite-plugin-dlight/CHANGELOG.md
@@ -1,5 +1,39 @@
 # vite-plugin-dlight
 
+## 1.0.0-beta.16
+
+### Patch Changes
+
+- Updated dependencies
+  - babel-preset-dlight@1.0.0-beta.16
+
+## 1.0.0-beta.15
+
+### Patch Changes
+
+- babel-preset-dlight@1.0.0-beta.15
+
+## 1.0.0-beta.14
+
+### Patch Changes
+
+- Updated dependencies
+  - babel-preset-dlight@1.0.0-beta.14
+
+## 1.0.0-beta.13
+
+### Patch Changes
+
+- Updated dependencies
+  - babel-preset-dlight@1.0.0-beta.13
+
+## 1.0.0-beta.12
+
+### Patch Changes
+
+- Updated dependencies
+  - babel-preset-dlight@1.0.0-beta.12
+
 ## 1.0.0-beta.11
 
 ### Patch Changes